worker/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
supabase==2.3.7
gunicorn==21.2.0
python-multipart==0.0.9
